Made site_url a property of the AccessToken
Working with the Salesforce API, you get the base url for API requests
back with your access token - and it may change depending on the
account the user logs in with. Since this needs to be saved and reloaded
with the access token, I've added that property and a new site_url
method that uses the property if set, or delegates to the Client
object (the original behaviour) if not.

I've also added the response content to the die message in request()
as this is useful for debugging. It may be necessary to make this
behaviour optional to shield the end user from the data, but the
result of a die shouldn't really make it to the end user anyway.